fix: cli invocation for Robyn
