+++ pyproject.toml
@@ -82,6 +82,7 @@ nox = "2023.4.22"
 websocket-client = "1.5.0"
 
 [tool.poetry.scripts]
+robyn = "robyn.cli:run"
 test_server = { callable = "integration_tests.base_routes:main" }
 
 [tool.ruff]
